Fixed compile breaking on Windows.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@1098 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/misc.cc
@@ -1948,7 +1948,7 @@ int place_branch(unsigned short place)
 {
     const unsigned branch = (unsigned) ((place >> 8) & 0xFF);
     const int lev = place & 0xFF;
-    return lev == 0xFF? -1 : branch;
+    return lev == 0xFF? -1 : (int) branch;
 }
 
 int place_depth(unsigned short place)
